+++ arch/x86/include/asm/agp.h
@@ -26,6 +26,6 @@
 #define alloc_gatt_pages(order)		\
 	((char *)__get_free_pages(GFP_KERNEL, (order)))
 #define free_gatt_pages(table, order)	\
-	free_pages((unsigned long)(table), (order))
+	free_pages((table), (order))
 
 #endif /* _ASM_X86_AGP_H */
